Fixed GRP display Bug
Display would freeze randomly as players were finishing up. I believe this was because a display list of size one was being created each time it was called messing with the stack. I fixed it by creating a nother list wich saves players as they finish at the front. this seems to be working well